Split C++ frontend into 3 files
